[BEAM-12009] Move CalcRelSplitter to rel subpackage.
